+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Release Notes
 Version 6.6
 -----------
  * More UI polish
+ * Share your collection
 
 Version 6.5.5
 -------------